Add a way to measure code coverage
[Code coverage](https://en.wikipedia.org/wiki/Code_coverage) is a
measure used to describe the degree to which the source code of a
program is executed when a particular test suite runs. A program with
high code coverage, measured as a percentage, has a lower chance of
containing undetected software bugs compared to a program with low code
coverage. For more information, see `README.md`.