Make the destructor of the Connection class non-virtual
A publicly inherited class should have a virtual or protected destructor
to prevent undefined behavior according to the [Core Guidelines][1]. The
`Connection` class is used as a base class, but it has no virtual
methods and the `ScopedConnection` class inherited from it has no member
variables of its own, so the destructor does not need to be virtual. And
not disconnecting the slot when the `ScopedConnection` is destroyed
using the base class pointer might actually be the desired behavior
(although this should preferably be done using the `release()` method).
And by specifying the destructor as non-virtual, we avoid creating a
virtual table and save a few bytes.

[1]: https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rc-dtor-virtual