Refactor and refresh unit tests
The unit tests are refactored and refreshed to improve readability and
maintainability. While doing this, a couple of issues were found: move
assigning a connected connection to another did not disconnect the
source, and move assigning a scoped connection to another did not first
disconnect the target connection. Both issues are now fixed.