Update ClangFormat config
The physical devices and media that were used to store and display code
in the past had a limited number of characters per line. Modern tools do
not have this limitation, but horizontal space is still a scarce and
valuable resource for readability. Therefore, some programmers and
projects choose to avoid indenting code within a namespace, among other
reasons and preferences.