Add combiner policy
Slots have supported arguments and return values since the first
implementation. However, it was not possible to send the return values
back to the signal caller. This change adds a combiner policy that
collects (and combines) the results of the slots into a single return
value for the caller. The default combiner returns the result of the
last slot called, but the combiner can be customized to return, for
example, the maximum value of any slot, the sum of all results, or even
all results in a container.

NOTE: A slot adapter should be implemented to skip over disconnected
slots during iteration. This would simplify the creation of combiner
policies. However, since this feature is not really used currently, it
can be deferred until a later time.